+++ src/index.ts
@@ -1,4 +1,4 @@
-export { isHex, isRgb, isRgba, isCmyk, isHsl, isColor } from './types/isType'
+export { isHex, isRgb, isRgba, isCmyk, isHsl, isHsla, isColor } from './types/isType'
 export { color2string, color2cssString } from './lib/color-models/color'
 export {
   hex2rgbOrRgba,
@@ -21,6 +21,7 @@ export {
 } from './lib/color-models/rgb'
 export { cmyk2rgb, cmyk2hex, cmyk2hsl, cmykString2Object } from './lib/color-models/cmyk'
 export { hsl2hex, hsl2rgb, hsl2cmyk, hslString2Object } from './lib/color-models/hsl'
+export { hslaToHex, hslaToRgb, hslaToRgba, hslaToHsl, hslaToCmyk } from './lib/color-models/hsla'
 export { getRandomColor } from './lib/misc/random'
 export { mix } from './lib/misc/mix'
 export { name } from './lib/color-names/name'