+++ src/lib/color-models/color.test.ts
@@ -10,8 +10,9 @@ test(`color2string`, () => {
   expect(color2string({ r: 0, g: 0, b: 0, a: 0 })).toBe('0, 0, 0, 0')
   expect(color2string({ c: 0, m: 0, y: 0, k: 0 })).toBe('0%, 0%, 0%, 0%')
   expect(color2string({ h: 0, s: 0, l: 0 })).toBe('0, 0%, 0%')
+  expect(color2string({ h: 0, s: 0, l: 0, a: 0 })).toBe('0, 0%, 0%, 0')
 
-  expect(() => color2string({ h: -1, s: 90, l: 50 })).toThrowError()
+  expect(() => color2string('#')).toThrowError()
 })
 
 ////////////////////////////////////////////////////////
@@ -24,6 +25,7 @@ test(`color2cssString`, () => {
   expect(color2cssString({ r: 0, g: 0, b: 0, a: 0 })).toBe('rgba(0, 0, 0, 0)')
   expect(color2cssString({ c: 0, m: 0, y: 0, k: 0 })).toBe('cmyk(0%, 0%, 0%, 0%)')
   expect(color2cssString({ h: 0, s: 0, l: 0 })).toBe('hsl(0, 0%, 0%)')
+  expect(color2cssString({ h: 0, s: 0, l: 0, a: 0 })).toBe('hsla(0, 0%, 0%, 0)')
 
-  expect(() => color2string({ h: -1, s: 90, l: 50 })).toThrowError()
+  expect(() => color2cssString('#')).toThrowError()
 })