+++ src/lib/color-models/color.ts
@@ -4,36 +4,26 @@ import { toUpper } from 'lodash'
 
 // Convert a color to a string format
 export function color2string(color: Color): string {
-  if (!isColor(color)) {
-    throw new Error(`${color} is not a color.`)
-  }
+  if (!isColor(color)) throw new Error(`${color} is not a color.`)
 
-  if (isHex(color)) {
-    return toUpper(color)
-  } else if (isRgb(color)) {
-    return `${color.r}, ${color.g}, ${color.b}`
-  } else if (isRgba(color)) {
-    return `${color.r}, ${color.g}, ${color.b}, ${color.a}`
-  } else if (isCmyk(color)) {
-    return `${color.c}%, ${color.m}%, ${color.y}%, ${color.k}%`
-  } else {
-    // is hsl
-    return `${color.h}, ${color.s}%, ${color.l}%`
-  }
+  if (isHex(color)) return toUpper(color)
+  else if (isRgb(color)) return `${color.r}, ${color.g}, ${color.b}`
+  else if (isRgba(color)) return `${color.r}, ${color.g}, ${color.b}, ${color.a}`
+  else if (isCmyk(color)) return `${color.c}%, ${color.m}%, ${color.y}%, ${color.k}%`
+  else if (isHsl(color)) return `${color.h}, ${color.s}%, ${color.l}%`
+  // @ts-ignore
+  else return `${color.h}, ${color.s}%, ${color.l}%, ${color.a}` // is hsla
 }
 
 // Convert a color to a string format usable in CSS
 export function color2cssString(color: Color): string {
-  if (isHex(color)) {
-    return toUpper(color)
-  } else if (isRgb(color)) {
-    return `rgb(${color.r}, ${color.g}, ${color.b})`
-  } else if (isRgba(color)) {
-    return `rgba(${color.r}, ${color.g}, ${color.b}, ${color.a})`
-  } else if (isCmyk(color)) {
-    return `cmyk(${color.c}%, ${color.m}%, ${color.y}%, ${color.k}%)`
-  } else {
-    // is hsl
-    return `hsl(${color.h}, ${color.s}%, ${color.l}%)`
-  }
+  if (!isColor(color)) throw new Error(`${color} is not a color.`)
+
+  if (isHex(color)) return toUpper(color)
+  else if (isRgb(color)) return `rgb(${color.r}, ${color.g}, ${color.b})`
+  else if (isRgba(color)) return `rgba(${color.r}, ${color.g}, ${color.b}, ${color.a})`
+  else if (isCmyk(color)) return `cmyk(${color.c}%, ${color.m}%, ${color.y}%, ${color.k}%)`
+  else if (isHsl(color)) return `hsl(${color.h}, ${color.s}%, ${color.l}%)`
+  // @ts-ignore
+  else return `hsla(${color.h}, ${color.s}%, ${color.l}%, ${color.a})` // is hsla
 }