+++ src/lib/color-models/rgb.ts
@@ -1,6 +1,6 @@
 import { round } from 'lodash'
 import { RGB, RGBA, CMYK, HEX, HSL, Color } from '../../types/types'
-import { isRgb, isRgba, isColor, isHex, isCmyk } from '../../types/isType'
+import { isRgb, isRgba, isHex, isCmyk } from '../../types/isType'
 import { applyFnToEachObjValue } from '../misc/utils'
 import { hex2rgba } from './hex'
 import { cmyk2rgb } from './cmyk'